# docs/swagger.yaml
basePath: /
definitions:
...
  helpers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Class:
    properties:
      created_at:
        type: string
      grade:
        type: integer
      id:
        type: integer
      name:
        type: string
      school_id:
        type: integer
      student_count:
        type: integer
    type: object
  models.School:
    properties:
      class_count:
        type: integer
      created_at:
        type: string
      director:
        type: string
      id:
        type: integer
      name:
        type: string
      student_count:
        type: integer
      user:
        $ref: '#/definitions/models.UserInfo'
      user_id:
        type: integer
    type: object
...
  models.UserInfo:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: База школ с ролями ROO и School.
  title: eduBase API
  version: "1.0"
paths:
...
  /classes:
    get:
      description: ROO — все классы, School — только свои
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Class'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список классов
      tags:
      - Classes
    post:
      consumes:
      - application/json
      description: School создаёт свой класс
      parameters:
      - description: Данные класса
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Class'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Class'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать новый класс
      tags:
      - Classes
  /classes/{id}:
    delete:
      description: School — только свои, ROO — любые
      parameters:
      - description: ID класса
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить класс
      tags:
      - Classes
    get:
      description: ROO — любой класс, School — только свой
      parameters:
      - description: ID класса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Class'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить класс по ID
      tags:
      - Classes
    put:
      consumes:
      - application/json
      description: ROO — может обновить любой, School — только свой
      parameters:
      - description: ID класса
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Class'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить класс
      tags:
      - Classes
...
  /school/me:
    get:
      description: Только для пользователей с ролью School
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.School'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить данные своей школы
      tags:
      - Schools
    put:
      consumes:
      - application/json
      description: Только для пользователей с ролью School
      parameters:
      - description: Поля для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.School'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить данные своей школы
      tags:
      - Schools
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"